+++ src/deploy_inst.h
@@ -21,8 +21,7 @@ typedef std::map<int, std::vector<std::string> > BuildSched;
 // lifetimes.
 class DeployInst : 
   public cyclus::Institution, 
-  public cyclus::toolkit::CommodityProducerManager,
-  public cyclus::toolkit::Position {
+  public cyclus::toolkit::CommodityProducerManager {
   #pragma cyclus note { \
     "doc": \
       "Builds and manages agents (facilities) according to a manually" \