+++ CHANGELOG.rst
@@ -30,7 +30,6 @@ Since last release
 * Updated Doxygen homepage (#632)
 
 **Fixed:**
-
 * Schedule Decommission in ``Reactor::Tick()`` instead of Decommission (#609)
 * When trades fail in Source due to packaging, send empty material instead of seg faulting (#629)
 * Logging of resource moves between ResBufs in Storage is INFO4 not INFO1 (#625)
@@ -38,7 +37,7 @@ Since last release
 * Update conributing guide to match current practice (#662)
 
 **Removed:**
-
+* Removed inheritance of toolkit position by Cycamore agents (#692)
 * Removed references to deprecated ``ResourceBuff`` class (#604)
 * Removed ``Libxml++`` from build requirements (#634)